+++ package.json
@@ -1,22 +1,22 @@
 {
-  "name": "cdrive-api1",
+  "name": "sf-cdn",
   "version": "0.0.0",
   "private": true,
   "scripts": {
     "start": "node app.js",
     "dev": "nodemon app.js",
-    "build:prod": "webpack"
+    "build": "webpack"
   },
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "~4.16.1",
     "formidable": "^1.2.2",
-    "mime-types": "^2.1.31",
-    "nodemon": "^2.0.12"
+    "mime-types": "^2.1.31"
   },
   "devDependencies": {
     "clean-webpack-plugin": "^4.0.0-alpha.0",
+    "nodemon": "^2.0.21",
     "webpack": "^5.45.1",
     "webpack-cli": "^4.7.2",
     "webpack-node-externals": "^3.0.0"
